Add the bash scripts used to create AWS resources for this POC
Creating a GitHub action to push the hub files to S3 required some
initial AWS setup. For now, the setup is codified in bash scripts
for future reference. However, once we prove that our S3 storage
is working as needed, we would want to think through the process
for Hubverse users to create their AWS resources.