Remove master from list of branch triggers
I had copied the [main, master] list from another workflow but
realized that the related AWS IAM role used by this workflow
will be limited to the main branch only. We may need to revisit
this if an older hub with a "master" branch wants to onboard,
but for now, let's be consistent between the workflow and the
AWS permissions.